requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
pyjwt==1.7.1
gunicorn==20.0.4
psycopg2==2.8.6
Flask-SQLAlchemy==2.4.4
